[fix] companionId를 Long에서 Int로 수정

+++ core/data/src/main/kotlin/com/tripmate/android/core/data/repository/TripListRepositoryImpl.kt
@@ -22,7 +22,7 @@ internal class TripListRepositoryImpl @Inject constructor(
         service.getParticipatedTripList(userId).data.map { it.toEntity() }
     }
 
-    override suspend fun selectMate(userId: Long, companionId: Long) = runSuspendCatching {
+    override suspend fun selectMate(userId: Long, companionId: Int) = runSuspendCatching {
         service.selectMate(MateSelectRequest(companionId, userId))
     }
 }
