+++ core/domain/src/main/kotlin/com/tripmate/android/domain/repository/TripListRepository.kt
@@ -6,5 +6,5 @@ import com.tripmate.android.domain.entity.triplist.ParticipatedCompanionInfoEnti
 interface TripListRepository {
     suspend fun getCreatedTripListByUser(): Result<List<CreatedCompanionInfoEntity>>
     suspend fun getTripsParticipatedByUser(): Result<List<ParticipatedCompanionInfoEntity>>
-    suspend fun selectMate(userId: Long, companionId: Long): Result<Unit>
+    suspend fun selectMate(userId: Long, companionId: Int): Result<Unit>
 }
