+++ core/network/src/main/kotlin/com/tripmate/android/core/network/response/ParticipatedTripListResponse.kt
@@ -16,7 +16,7 @@ data class ParticipatedTripListResponse(
 @Serializable
 data class ParticipatedCompanionInfo(
     @SerialName("companionId")
-    val companionId: Long,
+    val companionId: Int,
     @SerialName("title")
     val title: String,
     @SerialName("date")
