+++ feature/mate-review/src/main/kotlin/com/tripmate/android/mate_review/viewmodel/MateReviewViewModel.kt
@@ -26,7 +26,7 @@ class MateReviewViewModel @Inject constructor(
     private val mateRepository: MateRepository,
     savedStateHandle: SavedStateHandle,
 ) : ViewModel() {
-    private val companionId: Long = requireNotNull(savedStateHandle.get<Long>(COMPANION_ID))
+    private val companionId: Int = requireNotNull(savedStateHandle.get<Int>(COMPANION_ID))
     private val date: String = requireNotNull(savedStateHandle.get<String>(DATE))
     private val title: String = requireNotNull(savedStateHandle.get<String>(TITLE))
